+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 * Add vhost to access log
 * Add http/2 support using the `spdy` module
 * Add websocket support for http-only proxies
-* Add `Domain` model, where you can enable wildcard support
+* Add `Domain` model, needed for future wildcard support
 * `Site#registerHit` will now correctly see when a response ends, even with keep-alive enabled
 
 ## 0.2.0 (2017-09-01)